+++ src/service/PropertyService.ts
@@ -1,8 +1,6 @@
 import { IPropertyService } from '../interface/IPropertyService';
 import { IProperty } from '../interface/IProperty';
 import { IPropertyRepository } from '../interface/IPropertyRepository';
-
-
 export class PropertyService implements IPropertyService {
   private propertyRepository: IPropertyRepository;
 
@@ -24,11 +22,20 @@ export class PropertyService implements IPropertyService {
   }
 
   async updateProperty(id: number, propertyData: Partial<IProperty>): Promise<IProperty | null> {
+    const property = await this.propertyRepository.getPropertyById(id);
+    if(!property){
+      throw new Error(`Property with ID ${id} does not exist`);
+    }
+
     const updatedProperty = await this.propertyRepository.updateProperty(id, propertyData);
     return updatedProperty; 
   }
 
   async deleteProperty(id: number): Promise<boolean> {
+    const property = await this.propertyRepository.getPropertyById(id);
+    if(!property){
+      throw new Error(`Property with ID ${id} does not exist`);
+    }
     const deleted = await this.propertyRepository.deleteProperty(id);
     return deleted;
   }
